refactor: split streaming write RPC implementation
