+++ generator/integration_tests/golden/internal/golden_kitchen_sink_auth_decorator.cc
@@ -20,6 +20,7 @@
 #include "google/cloud/internal/async_read_write_stream_auth.h"
 #include "google/cloud/internal/async_streaming_read_rpc_auth.h"
 #include "google/cloud/internal/async_streaming_write_rpc_auth.h"
+#include "google/cloud/internal/streaming_write_rpc_impl.h"
 #include <generator/integration_tests/test.grpc.pb.h>
 #include <memory>
