+++ google/cloud/google_cloud_cpp_grpc_utils.bzl
@@ -68,6 +68,7 @@ google_cloud_cpp_grpc_utils_hdrs = [
     "internal/streaming_read_rpc.h",
     "internal/streaming_read_rpc_logging.h",
     "internal/streaming_write_rpc.h",
+    "internal/streaming_write_rpc_impl.h",
     "internal/streaming_write_rpc_logging.h",
     "internal/time_utils.h",
     "internal/unified_grpc_credentials.h",
@@ -94,7 +95,7 @@ google_cloud_cpp_grpc_utils_srcs = [
     "internal/populate_grpc_options.cc",
     "internal/retry_loop_helpers.cc",
     "internal/streaming_read_rpc.cc",
-    "internal/streaming_write_rpc.cc",
+    "internal/streaming_write_rpc_impl.cc",
     "internal/time_utils.cc",
     "internal/unified_grpc_credentials.cc",
 ]
