+++ google/cloud/google_cloud_cpp_grpc_utils.cmake
@@ -87,8 +87,9 @@ add_library(
     internal/streaming_read_rpc.cc
     internal/streaming_read_rpc.h
     internal/streaming_read_rpc_logging.h
-    internal/streaming_write_rpc.cc
     internal/streaming_write_rpc.h
+    internal/streaming_write_rpc_impl.cc
+    internal/streaming_write_rpc_impl.h
     internal/streaming_write_rpc_logging.h
     internal/time_utils.cc
     internal/time_utils.h
