+++ google/cloud/internal/streaming_write_rpc_impl.cc
