+++ google/cloud/storage/internal/storage_round_robin_test.cc
@@ -16,6 +16,7 @@
 #include "google/cloud/storage/testing/mock_storage_stub.h"
 #include "google/cloud/internal/async_streaming_read_rpc_impl.h"
 #include "google/cloud/internal/async_streaming_write_rpc_impl.h"
+#include "google/cloud/internal/streaming_write_rpc_impl.h"
 #include "google/cloud/testing_util/status_matchers.h"
 #include "absl/memory/memory.h"
 #include <gmock/gmock.h>
